Refactor: socket 통신 부분 커스텀 hook 화로 분리
- ChatbotPage에서 비지니스 로직을 감추고 동작과 view만 보이기위에 명시성을 부여하기 위해 분리
- 자질구레한 로직이 길어지면 custom hook화를 한다고 보시면 되겠습니다.
